Removed redundant table creation logic and abstract the instrumentation framework to an inherited class
